feat: implement SpoilerCapable interface and enhance spoiler functionality in markdown components
