+++ android/src/main/java/com/swmansion/enriched/markdown/spoiler/SegmentKey.kt
@@ -4,3 +4,10 @@ data class SegmentKey(
   val spanIdentity: Int,
   val line: Int,
 )
+
+data class SegmentRect(
+  val left: Float,
+  val top: Float,
+  val width: Float,
+  val height: Float,
+)
