+++ android/src/main/java/com/swmansion/enriched/markdown/spoiler/SpoilerCapable.kt
@@ -1,0 +1,5 @@
+package com.swmansion.enriched.markdown.spoiler
+
+interface SpoilerCapable {
+  val spoilerOverlayDrawer: SpoilerOverlayDrawer?
+}
