+++ android/src/main/java/com/swmansion/enriched/markdown/spoiler/SpoilerConstants.kt
@@ -1,0 +1,3 @@
+package com.swmansion.enriched.markdown.spoiler
+
+internal const val REVEAL_DURATION_MS = 450L
