+++ android/src/main/java/com/swmansion/enriched/markdown/spoiler/SpoilerOverlayDrawer.kt
@@ -1,8 +1,6 @@
 package com.swmansion.enriched.markdown.spoiler
 
 import android.graphics.Canvas
-import android.graphics.Color
-import android.graphics.Paint
 import android.text.Spanned
 import android.widget.TextView
 import com.swmansion.enriched.markdown.spans.SpoilerSpan
@@ -150,47 +148,5 @@ class SpoilerOverlayDrawer(
       existing?.stop()
       return null
     }
-
-    internal fun computeSegmentRect(
-      layout: android.text.Layout,
-      line: Int,
-      segmentStart: Int,
-      segmentEnd: Int,
-      fontMetrics: Paint.FontMetrics,
-      paddingLeft: Float,
-      paddingTop: Float,
-    ): SegmentRect? {
-      val startHorizontal = layout.getPrimaryHorizontal(segmentStart)
-      val endHorizontal =
-        if (segmentEnd >= layout.getLineEnd(line)) {
-          layout.getLineRight(line)
-        } else {
-          layout.getPrimaryHorizontal(segmentEnd)
-        }
-      val baseline = layout.getLineBaseline(line).toFloat()
-
-      val left = minOf(startHorizontal, endHorizontal) + paddingLeft
-      val right = maxOf(startHorizontal, endHorizontal) + paddingLeft
-      val top = baseline + fontMetrics.ascent + paddingTop
-      val bottom = baseline + fontMetrics.descent + paddingTop
-      val width = right - left
-      val height = bottom - top
-      return if (width > 0 && height > 0) SegmentRect(left, top, width, height) else null
-    }
-
-    internal fun colorWithAlpha(
-      color: Int,
-      alpha: Float,
-    ): Int {
-      val alphaComponent = (Color.alpha(color) * alpha).toInt().coerceIn(0, 255)
-      return Color.argb(alphaComponent, Color.red(color), Color.green(color), Color.blue(color))
-    }
   }
 }
-
-data class SegmentRect(
-  val left: Float,
-  val top: Float,
-  val width: Float,
-  val height: Float,
-)
